% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_populations.R
\name{autocorr_estim}
\alias{autocorr_estim}
\title{Estimate autocorrelation in vital rates for a list of data frames.}
\usage{
autocorr_estim(listdf, groupvars = NULL, survival_col = "est_surv",
  fertility_col = "est_fecund")
}
\arguments{
\item{listdf}{The list of data frames.}

\item{groupvars}{The variables you want to group by (e.g., timesteps).}

\item{survival_col}{The name of the survival variable. The default value is
'est_surv', which is what the variable is called in the output of
\code{\link{autocorr_sim}}.}

\item{fertility_col}{The name of the fertility variable. The default value is
'est_fecund', which is what the variable is called in the output of
\code{\link{autocorr_sim}}.}
}
\value{
A data frame with one column for the sample autocorrelation of the
  survival, one for the sample autocorrelation of the fertility, and one
  column for each grouping variable specified in the function call.
}
\description{
This function estimates the autocorrelation in vital rates (survival and
fertility) for a list of data frames with a variable of estimated survival
and fertility, e.g. the output of \code{\link{autocorr_sim}}, though you can
use it on any list of data frames with a similar structure. You can also
group the autocorrelation estimates by any variable that occurs in all the
data frames in the list.
}
\examples{
survival_range <- autocorr_sim(timesteps = 30, start = 200, phi = 0.3,
                               survMean = c(0.1, 0.2, 0.3, 0.4, 0.5), survSd = 0.5,
                               fecundMean = 0.6, fecundSd = 0.5, replicates = 50)
estimates <- autocorr_estim(survival_range, groupvars = 'survMean')
head(estimates)
}
