% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colored_multi_rnorm}
\alias{colored_multi_rnorm}
\title{Generate Multiple Cross-Correlated & Autocorrelated Variables}
\usage{
colored_multi_rnorm(timesteps, mu, sigma, phi, corrMatrix)
}
\arguments{
\item{timesteps}{The number of temporally autocorrelated random numbers (one
per timestep) you want.}

\item{mu}{A vector giving the mean of each variable.}

\item{sigma}{A vector giving the standard deviation of each variable.}

\item{phi}{A vector giving the temporal autocorrelation of each variable.}

\item{corrMatrix}{A valid correlation matrix. The number of rows/columns must match the length of the mu, sigma, and phi vectors.}
}
\value{
A matrix with as many rows as timesteps and as many columns as mu/sigma/phi values.
}
\description{
Generates random variables that are correlated to each other and temporally autocorrelated.
}
\examples{
corr <- matrix(c(1, 0.53, 0.73, 0.53, 1, 0.44, 0.73, 0.44, 1), nrow = 3)
test <- colored_multi_rnorm(100, c(0, 3, 5), c(1, 0.5, 1), c(0.5, -0.3, 0), corr)
var(test)
library(dplyr)
test \%>\% as.data.frame() \%>\% summarize_all(.funs = c("mean", "sd", "autocorrelation"))
}
