% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorednoise.R
\docType{package}
\name{colorednoise}
\alias{colorednoise}
\alias{colorednoise-package}
\title{\code{colorednoise} package}
\description{
Simulate Temporally Autocorrelated Populations
}
\details{
See the README on
\href{https://cran.r-project.org/package=colorednoise/README.html}{CRAN}
or \href{https://github.com/japilo/colorednoise#readme}{GitHub}
}
