% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_populations.R
\name{matrix_model}
\alias{matrix_model}
\title{Temporally Autocorrelated Matrix Models}
\usage{
matrix_model(meanMat, sdMat, phiMat, initialPop, timesteps)
}
\arguments{
\item{meanMat}{A standard Leslie matrix with mean values for each element.}

\item{sdMat}{A standard Leslie matrix with standard deviations for each element.}

\item{phiMat}{A standard Leslie matrix with a temporal autocorrelation value for each element.}

\item{initialPop}{An initial population vector. The length must be the same as the number of classes in the matrices.}

\item{timesteps}{The number of timesteps you would like to simulate the population.}
}
\value{
A data frame with a timestep column and a column for each class with population sizes at each timestep.
}
\description{
Simulate a structured population with temporal autocorrelation using standard Leslie matrices.
Each element in the Leslie matrix has a specified mean, variance, and temporal autocorrelation value.

The matrix can have arbitrary dimensions and can have transitions besides linear survival. This model
includes environmental stochasticity with colored noise but not demographic stochasticity.
}
\examples{
meanMat <- matrix(c(0.55, 0.6, 0.24, 0.4), byrow = T, ncol = 2)
sdMat <- matrix(c(0.3, 0.35, 0.05, 0.1), byrow = T, ncol = 2)
phiMat <- matrix(c(-0.2, -0.2, 0, 0), byrow = T, ncol = 2)
initialPop <- c(100, 100)
sim <- matrix_model(meanMat, sdMat, phiMat, initialPop, 50)
head(sim)
}
